refactor(DEV-11737): rename `currenices.ts` to `currencies.ts (fix typo)
