+++ packages/sdk-react/src/core/utils/currencies.ts
