+++ packages/sdk-react/src/core/utils/index.ts
@@ -1,2 +1,2 @@
 export * from './countries';
-export * from './currenices';
+export * from './currencies';
